Add disable_ipv6 state
For some reason, Comcast drops my packets for times as often if I use
IPv6. This state is obviously made for to solve a specific problem that
is likely to only apply to my currently living situation. As such it is
important to make sure that this repository doesn't push this state to
all minions once this repository has an organized top file.