Add a symbolic link to dotfiles pillar
Every state that may use a pillar should include an example of such a
pillar. In cases where the pillar has no security constraints, then the
actual pillar should be included in this repository. If the example
pillar would be the same as the actual pillar, then a symlink to the
pillar should be included in the state itself. This makes it easy to
find the example, and keeps things DRY.