Break python into several states with a meta state
In general I would like each folder's init.sls to just include all other
states in the folder. This leads to predictability in the layout as well
as a transitive property in which including state ensures that all
substates will also be included. Practically speaking this means that by
including a folder, you are including all states in that folder, and all
subfolders. This means you don't have to worry about missing a state
buried too far along the tree.

Also, I am introducing a convention where I will name a state 'basic' if
it is the core state for the folder. Since python/init.sls is a meta
state, python must be installed somewhere other than the init state,
i.e., basic.