Remove RDO state
I haven't gotten RDO to work properly for me and this state is more a
liability than anything else. Firstly, it introduces and inconsistency
because I have not been including it in the init file for
virtualization. Secondly, accidentally running the state would be quite
unfortunate, hence it not being in the init file. If I really want to, I
can dig through git's history to find this in the future.