Delete minecraft state because I do not use it
Furthermore, the state relied on downloading the jar with a particular
hash. I am not sure, but I bet this would change with new releases. With
the frequency that I play Minecraft, that would have made the state
worthless. Before the commit, it was just making it harder to find what
I wanted.