Remove users state because it was doing more harm than good
The problem with the state is that it relies on secure information in a
pillar which I would not be able to store in the repository with
everything else. If I forgot to take the time out to make that pillar, I
run the risk of having my password changed to the extremely insecure
default. For now, I will just manage this by hand, as much as it pains
me to say that about anything.