Remove 'root' from users pillar
Configuring root's home directory is probably a bad idea. This was
included originally because I needed to do a lot of work as root and it
was more convenient to have my current configuration across my normal
and root account. However, I have decided to stop configuring root's
home directory and to instead use 'sudo -s' to switch to root with my
configuration.